changed type from float to double

+++ src/constants.hpp
@@ -1,15 +1,15 @@
 # ifndef CONSTANTS_HPP 
 # define CONSTANTS_HPP 
 
-namespace constants
+namespace constants 
 {
     constexpr int n_cohort = 4; 
     
-    constexpr float p[n_cohort] = {
-        0.15f, 
-        0.10f, 
-        0.30f, 
-        0.45f
+    constexpr double p[n_cohort] = {
+        0.15, 
+        0.10, 
+        0.30, 
+        0.45
     };
 
     constexpr int c[n_cohort] = {
@@ -19,11 +19,11 @@ namespace constants
         300
     }; 
 
-    constexpr float v[n_cohort] = {
-        2.0f, 
-        1.0f, 
-        0.4f, 
-        0.333f
+    constexpr double v[n_cohort] = {
+        2.0, 
+        1.0, 
+        0.4, 
+        0.333
     }; 
 }
