+++ src/multinomial.hpp
@@ -10,9 +10,9 @@ class Multinomial
 public: 
     int m_total; 
     std::array<int, constants::n_cohort> m_array; 
-    float m_density; 
+    double m_density; 
     int m_entries; 
-    float m_value; 
+    double m_value; 
     Multinomial(int n1, int n2, int n3, int n4); 
 };
 
@@ -20,6 +20,6 @@ public:
 auto multinomial_pdf( 
     const int& n, 
     const std::array<int, constants::n_cohort>& n_array
-) -> float; 
+) -> double; 
 
 #endif
